[networks] Add unit tests for `utils.go`
