+++ pkg/network/telemetry/metric_test.go
@@ -89,29 +89,3 @@ func TestMetricOperations(t *testing.T) {
 		assert.Equal(int64(15), m1.Get())
 	})
 }
-
-func TestSplitTagsAndOpts(t *testing.T) {
-	assert := assert.New(t)
-
-	t.Run("only tags", func(t *testing.T) {
-		tags, opts := splitTagsAndOptions([]string{"tag:a", "tag:c", "tag:b"})
-		assert.Equal([]string{"tag:a", "tag:b", "tag:c"}, tags)
-		assert.Len(opts, 0)
-	})
-
-	t.Run("only opts", func(t *testing.T) {
-		tags, opts := splitTagsAndOptions([]string{"_opt3", "_opt2", "_opt1"})
-		assert.Equal([]string{"_opt1", "_opt2", "_opt3"}, opts)
-		assert.Len(tags, 0)
-	})
-
-	t.Run("tags and opts", func(t *testing.T) {
-		tags, opts := splitTagsAndOptions(
-			[]string{"_opt3", "tag:a", "_opt2", "tag:b", "_opt1", "tag:c"},
-		)
-
-		assert.Equal([]string{"tag:a", "tag:b", "tag:c"}, tags)
-		assert.Equal([]string{"_opt1", "_opt2", "_opt3"}, opts)
-	})
-
-}
